requirements.txt
langchain==1.2.0
langchain-community==0.4.1
langchain-openai==1.1.6
opensearch-py==3.1.0
langgraph==1.0.5
streamlit==1.52.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
pyodbc==5.3.0
pytest==9.0.2
black==25.12.0
pandas==2.3.3
openpyxl==3.1.5
